refactor: ProductsDetailList에 필요한 extenstion들 수정
